[modify] split permit params to its own module
